Refactor `ui_fsmenu/addons.cc`
