+++ src/ui_fsmenu/CMakeLists.txt
@@ -2,12 +2,24 @@ wl_library(ui_fsmenu
   SRCS
     about.cc
     about.h
-    addons.cc
-    addons.h
-    addons_packager.cc
-    addons_packager.h
-    addons_packager_box.cc
-    addons_packager_box.h
+    addons/contact.cc
+    addons/contact.h
+    addons/login_box.cc
+    addons/login_box.h
+    addons/manager.cc
+    addons/manager.h
+    addons/packager.cc
+    addons/packager.h
+    addons/packager_box.cc
+    addons/packager_box.h
+    addons/progress.cc
+    addons/progress.h
+    addons/remote_interaction.cc
+    addons/remote_interaction.h
+    addons/rows_ui.cc
+    addons/rows_ui.h
+    addons/screenshot_upload.cc
+    addons/screenshot_upload.h
     campaigndetails.cc
     campaigndetails.h
     campaigns.cc
